refactor(protocol): structure AccountRegistry files into epoch dirs
